fix logfiles logic

+++ src/main/logs.ts
@@ -26,11 +26,13 @@ export function setupLogs(
   holochainLogsToTerminal: boolean,
 ) {
   const logFilePath = path.join(launcherFileSystem.appLogsDir, 'we.log');
-  const stats = fs.statSync(logFilePath);
-  // If logfile is larger than 1GB, delete it
-  if (stats.size > 1e9) {
-    console.log("Found a log file that's larger than 1GB. Deleting it.");
-    fs.rmSync(logFilePath);
+  if (fs.existsSync(logFilePath)) {
+    const stats = fs.statSync(logFilePath);
+    // If existing logfile is larger than 1GB, delete it
+    if (stats.size > 1e9) {
+      console.log("Found a log file that's larger than 1GB. Deleting it.");
+      fs.rmSync(logFilePath);
+    }
   }
 
   // Use log file rotation with max size of a single file of 50MB and max 5 total files
@@ -39,7 +41,7 @@ export function setupLogs(
     maxsize: 50_000_000,
     maxFiles: 5,
   });
-  // const logFileTransport = new transports.File({ filename: logFilePath });
+
   const lairLogger = createLairLogger(logFileTransport);
 
   weEmitter.on(LAIR_LOG, (log) => {
